Refactoring structure and naming
