+++ Front/carrito-app/src/app/pages/whises/card/card.component.css
