+++ Front/carrito-app/src/app/shared/barra/barra.component.css
