+++ Front/carrito-app/src/app/shared/footer/footer.component.spec.ts
