--- Front/carrito-app/src/app/whises/wishes.spec.ts
@@ -1,7 +1,0 @@
-import { Curso } from './wish';
-
-describe('Curso', () => {
-  it('should create an instance', () => {
-    expect(new Curso()).toBeTruthy();
-  });
-});
